Optimize request latency
In theory, the same query made with in client mode vs batch mode should be faster
in client mode, since it doesn't have to load the data files. However, for small
requests, Nagle's algorithm introduces a large delay. At least for the test cases
in this directory, the batch query takes less than 100ms while the client query
takes 400ms!

This commit sets TCP_NODELAY, to flush queries immediately. As a results, for the
case I was testing the timing dropped to 16ms.

This commit also includes code to run the server without daemonizing, but I still need
to add a command line for it.