Add a thread pool to serve requests
Turns out it is really simple to add a thread pool for boost asio based
servers.

I tried to test it by modifying the client to loop 10000 times and then running 4
clients at the same time. However, the client time is dominated by the DNS query and
other systems stuff, so my test to prove the benefit of a thread pool was inconclusive.
I think I should come up with a query that is really slow, so as to dwarf DNS lookup and
other network overhead.

However, this test at least seems to demonstrate that it's safe to unleash a bunch of threads
on this code without synchronisation. It seems that my assumption that the data structures
are read-only is confirmed.